% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{quiz}
\alias{quiz}
\title{Create a multiple choice quiz}
\usage{
quiz(question, correct_answer, answer_info, answers)
}
\description{
Provide a question, answers, select the correct answer, and provide
information on incorrect answer
}
